+++ GSVDevice/GSV6_BasicFrameType.py
@@ -46,8 +46,8 @@ __author__ = 'Dennis Rump'
 #
 ###############################################################################
 
-import GSV_Exceptions
-from GSV6_ErrorCodes import error_code_to_error_shortcut
+from GSVDevice import GSV_Exceptions
+from GSVDevice.GSV6_ErrorCodes import error_code_to_error_shortcut
 
 
 class BasicFrame:
