+++ GSVDevice/GSV6_ErrorCodes.py
