+++ GSVDevice/GSV_Exceptions.py
