+++ GSVDevice/__init__.py
@@ -1,0 +1,2 @@
+from .gsv8 import gsv8
+from .GSVDeviceSerial import GSVDeviceSerial
